normalize xsb route layouts
