+++ packages/esuite/app/(beta)/xsb/generations/layout.tsx
@@ -1,14 +1,3 @@
-import { appConfig } from '@/config/config'
-import type { Metadata } from 'next'
-
-export const metadata: Metadata = {
-  title: {
-    template: `${appConfig.siteTitle} · Generations · %s`,
-    default: 'Generations',
-  },
-  description: appConfig.siteDescription,
-}
-
 export default function Layout({ children }: { children: React.ReactNode }) {
   return <div className="flex h-dvh">{children}</div>
 }
