--- packages/esuite/app/(beta)/xsb/patterns/[[...patternId]]/page.tsx
@@ -1,5 +1,0 @@
-import { PatternEditorPage } from '@/components/patterns/PatternEditor'
-
-export default function Page({ params }: { params: { patternId?: string[] } }) {
-  return <PatternEditorPage patternId={params.patternId?.[0]} />
-}
