+++ packages/esuite/app/(beta)/xsb/patterns/page.tsx
@@ -1,0 +1,5 @@
+import { PatternEditorPage } from '@/components/patterns/PatternEditor'
+
+export default function Page() {
+  return <PatternEditorPage />
+}
